requirements.txt
langchain
langchain-community
langchain-groq
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-text-splitters
langchain-core
streamlit
faiss-cpu
sentence-transformers
pypdf
python-dotenv
